consolidate sharing e2e specs
